load fonts from file system
