+++ _api/frames/router.tsx
@@ -1,6 +1,4 @@
 /* eslint-disable no-console */
-import { readFile } from 'node:fs/promises';
-import { join } from 'node:path';
 import { Readable } from 'node:stream';
 import { ReadableStream } from 'node:stream/web';
 import React from 'react';
@@ -9,9 +7,9 @@ import { VercelRequest, VercelResponse } from '@vercel/node';
 import { ImageResponse } from '@vercel/og';
 import { Path } from 'path-parser';
 
-import { IS_LOCAL_ENV } from '../../api-lib/config';
 import { webAppURL } from '../../src/config/webAppURL';
 
+import { loadFonts } from './_fonts.ts';
 import { ErrorFrame } from './ErrorFrame';
 import { RenderFrameMeta } from './FrameMeta';
 import { FramePostInfo, getFramePostInfo } from './getFramePostInfo';
@@ -28,16 +26,6 @@ import { PersonaZeroFrame } from './personas/PersonaZeroFrame';
 
 export const FRAME_ROUTER_URL_BASE = `${webAppURL('colinks')}/api/frames/router`;
 
-declare type Weight = 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900;
-declare type Style = 'normal' | 'italic';
-interface FontOptions {
-  data: Buffer | ArrayBuffer;
-  name: string;
-  weight?: Weight;
-  style?: Style;
-  lang?: string;
-}
-
 type PathWithHandler = {
   path: Path;
   handler: (
@@ -48,30 +36,13 @@ type PathWithHandler = {
   method: 'GET' | 'POST';
 };
 
+//load the fonts just once, not once per handler
+const fonts = await loadFonts();
+
 const router: {
   paths: PathWithHandler[];
 } = {
   paths: [],
-};
-
-const getPath = (name: string) =>
-  join(process.cwd(), 'public', 'fonts', `${name}.ttf`);
-const createFont = async (name: string, file: string) => {
-  // TODO: fix font loading in vercel, url fetching is very slow
-
-  let fontData: ArrayBuffer;
-  if (IS_LOCAL_ENV) {
-    fontData = await readFile(getPath(file));
-  } else {
-    const baseUrl = webAppURL('colinks');
-    const path = new URL(`${baseUrl}/fonts/${file}.ttf`);
-    fontData = await fetch(path).then(res => res.arrayBuffer());
-  }
-
-  return {
-    name: name,
-    data: fontData,
-  };
 };
 
 export default async function (req: VercelRequest, res: VercelResponse) {
@@ -207,35 +178,6 @@ const addFrame = (frame: Frame) => {
     }
   );
 
-  const loadFonts = async (): Promise<FontOptions[]> => {
-    const startTime = Date.now();
-    const fonts = await Promise.all([
-      {
-        ...(await createFont('Denim', 'Denim-Regular')),
-        weight: 400,
-        style: 'normal',
-      },
-      {
-        ...(await createFont('Denim', 'Denim-RegularItalic')),
-        weight: 400,
-        style: 'italic',
-      },
-      {
-        ...(await createFont('Denim', 'Denim-SemiBold')),
-        weight: 600,
-        style: 'normal',
-      },
-      {
-        ...(await createFont('Denim', 'Denim-SemiBoldItalic')),
-        weight: 600,
-        style: 'italic',
-      },
-    ]);
-    const endTime = Date.now();
-    console.log('Font load time:', endTime - startTime, 'ms');
-    return fonts as FontOptions[];
-  };
-
   // always add an image route
   addPath(
     `/img/${frame.id}${frame.resourceIdentifier.resourcePathExpression}`,
@@ -245,7 +187,7 @@ const addFrame = (frame: Frame) => {
         // debug: true,
         height: 1000,
         width: 1000,
-        fonts: await loadFonts(),
+        fonts,
       });
       // no cache
       //
